feat: add compute proof/query checking to sdk run fn
