+++ circuit/src/tests/keccak.rs
@@ -1,6 +1,6 @@
 use std::sync::{Arc, Mutex};
 
-use axiom_codec::HiLo;
+use axiom_codec::{constants::USER_MAX_OUTPUTS, HiLo};
 use axiom_query::{
     axiom_eth::{
         halo2_base::{
@@ -39,7 +39,7 @@ use crate::{
     subquery::caller::SubqueryCaller,
     tests::utils::{account_call, header_call, EmptyCircuitInput},
     types::AxiomCircuitParams,
-    utils::get_provider,
+    utils::{check_compute_proof_format, check_compute_query_format, get_provider},
     witness,
 };
 
@@ -191,6 +191,7 @@ pub fn test_compute_query<S: AxiomCircuitScaffold<Http, Fr>>(_circuit: S) {
         final_output.clone(),
         AxiomCircuitParams::Base(circuit.builder.config_params),
         agg_vk,
+        USER_MAX_OUTPUTS,
     );
     // TEMP
     let kzg_params = gen_srs(agg_circuit_params.degree);
